fixed frameworks and tests
