+++ NetworkKit.xcodeproj/xcuserdata/alextelek.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -7,42 +7,27 @@
 		<key>NetworkKit.xcscheme_^#shared#^_</key>
 		<dict>
 			<key>orderHint</key>
-			<integer>0</integer>
+			<integer>1</integer>
 		</dict>
 		<key>NetworkKitExample.xcscheme_^#shared#^_</key>
 		<dict>
 			<key>orderHint</key>
-			<integer>2</integer>
+			<integer>0</integer>
 		</dict>
 		<key>NetworkKitMac.xcscheme_^#shared#^_</key>
 		<dict>
 			<key>orderHint</key>
 			<integer>3</integer>
 		</dict>
-		<key>NetworkKitMacTests.xcscheme_^#shared#^_</key>
-		<dict>
-			<key>orderHint</key>
-			<integer>6</integer>
-		</dict>
-		<key>NetworkKitTests.xcscheme_^#shared#^_</key>
-		<dict>
-			<key>orderHint</key>
-			<integer>1</integer>
-		</dict>
 		<key>NetworkKitTvOS.xcscheme_^#shared#^_</key>
 		<dict>
 			<key>orderHint</key>
 			<integer>5</integer>
 		</dict>
-		<key>NetworkKitTvOSTests.xcscheme_^#shared#^_</key>
-		<dict>
-			<key>orderHint</key>
-			<integer>7</integer>
-		</dict>
 		<key>NetworkKitWatchOS.xcscheme_^#shared#^_</key>
 		<dict>
 			<key>orderHint</key>
-			<integer>4</integer>
+			<integer>7</integer>
 		</dict>
 	</dict>
 	<key>SuppressBuildableAutocreation</key>
@@ -78,6 +63,11 @@
 			<true/>
 		</dict>
 		<key>2BD3D3C91C92EE920080BC62</key>
+		<dict>
+			<key>primary</key>
+			<true/>
+		</dict>
+		<key>2BF2B6EF1CA44B7F0003939B</key>
 		<dict>
 			<key>primary</key>
 			<true/>
