+++ NetworkKit/Info.plist
@@ -19,7 +19,7 @@
 	<key>CFBundleSignature</key>
 	<string>????</string>
 	<key>CFBundleVersion</key>
-	<string>171</string>
+	<string>289</string>
 	<key>NSPrincipalClass</key>
 	<string></string>
 </dict>
