--- NetworkKitMac/NetworkKitMac.h
@@ -1,19 +1,0 @@
-//
-//  NetworkKitMac.h
-//  NetworkKitMac
-//
-//  Created by Alex Telek on 17/03/2016.
-//  Copyright © 2016 alextelek. All rights reserved.
-//
-
-#import <Cocoa/Cocoa.h>
-
-//! Project version number for NetworkKitMac.
-FOUNDATION_EXPORT double NetworkKitMacVersionNumber;
-
-//! Project version string for NetworkKitMac.
-FOUNDATION_EXPORT const unsigned char NetworkKitMacVersionString[];
-
-// In this header, you should import all the public headers of your framework using statements like #import <NetworkKitMac/PublicHeader.h>
-
-
