--- NetworkKitTvOS/NetworkKitTvOS.h
@@ -1,19 +1,0 @@
-//
-//  NetworkKitTvOS.h
-//  NetworkKitTvOS
-//
-//  Created by Alex Telek on 17/03/2016.
-//  Copyright © 2016 alextelek. All rights reserved.
-//
-
-#import <UIKit/UIKit.h>
-
-//! Project version number for NetworkKitTvOS.
-FOUNDATION_EXPORT double NetworkKitTvOSVersionNumber;
-
-//! Project version string for NetworkKitTvOS.
-FOUNDATION_EXPORT const unsigned char NetworkKitTvOSVersionString[];
-
-// In this header, you should import all the public headers of your framework using statements like #import <NetworkKitTvOS/PublicHeader.h>
-
-
