--- NetworkKitTvOSTests/NetworkKitTvOSTests.swift
@@ -1,35 +1,0 @@
-//
-//  NetworkKitTvOSTests.swift
-//  NetworkKitTvOSTests
-//
-//  Created by Alex Telek on 20/03/2016.
-//  Copyright © 2016 alextelek. All rights reserved.
-//
-
-import XCTest
-
-class NetworkKitTvOSTests: XCTestCase {
-    
-    override func setUp() {
-        super.setUp()
-        // Put setup code here. This method is called before the invocation of each test method in the class.
-    }
-    
-    override func tearDown() {
-        // Put teardown code here. This method is called after the invocation of each test method in the class.
-        super.tearDown()
-    }
-    
-    func testExample() {
-        // This is an example of a functional test case.
-        // Use XCTAssert and related functions to verify your tests produce the correct results.
-    }
-    
-    func testPerformanceExample() {
-        // This is an example of a performance test case.
-        self.measureBlock {
-            // Put the code you want to measure the time of here.
-        }
-    }
-    
-}
