--- NetworkKitWatchOS/NetworkKitWatchOS.h
@@ -1,19 +1,0 @@
-//
-//  NetworkKitWatchOS.h
-//  NetworkKitWatchOS
-//
-//  Created by Alex Telek on 17/03/2016.
-//  Copyright © 2016 alextelek. All rights reserved.
-//
-
-#import <WatchKit/WatchKit.h>
-
-//! Project version number for NetworkKitWatchOS.
-FOUNDATION_EXPORT double NetworkKitWatchOSVersionNumber;
-
-//! Project version string for NetworkKitWatchOS.
-FOUNDATION_EXPORT const unsigned char NetworkKitWatchOSVersionString[];
-
-// In this header, you should import all the public headers of your framework using statements like #import <NetworkKitWatchOS/PublicHeader.h>
-
-
